requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
captcha==0.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
cssselect2==0.7.0
Django==3.2
django-phonenumber-field==7.0.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-registration-redux==2.12
django-simple-captcha==0.5.17
django-xhtml2pdf==0.0.4
django_debug_toolbar==3.8.1
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
phonenumbers==8.13.7
Pillow==9.4.0
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.6.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.9
